bugfix: 1.新增根据 clusterPhyId 获取物理集群配置信息；2。根据 gatewayClusterId 获取 gateway 配置信息；3。修复zeus查看任务执行完成后的标准输出接口；4。插件列表新增配置信息接口

+++ arius-admin/arius-admin-core/src/main/java/com/didichuxing/datachannel/arius/admin/core/service/gateway/impl/GatewayNodeServiceImpl.java
@@ -93,6 +93,9 @@ public class GatewayNodeServiceImpl implements GatewayNodeService {
 		
 		@Override
 		public Boolean deleteBatch(List<Integer> ids) {
-				return gatewayClusterNodeDAO.deleteBatch(ids)==ids.size();
+				if (CollectionUtils.isEmpty(ids)) {
+						return true;
+				}
+				return gatewayClusterNodeDAO.deleteBatch(ids) == ids.size();
 		}
 }
